refactor: restructure NavSection components
