+++ src/components/organisms/NewNavigatorPane/NavSectionRenderer/index.ts
@@ -1,0 +1,1 @@
+export {default} from './NavSectionRenderer';
