+++ src/components/organisms/NewNavigatorPane/NavSectionRenderer/styled.tsx
